ApiDiff: Move TryGetRecordConstructor from GenAPI.INamedTypeSymbolExtensions up to ApiSymbolExtensions.SymbolExtensions (#45801)
