+++ src/Compatibility/GenAPI/Microsoft.DotNet.GenAPI/INamedTypeSymbolExtensions.cs
@@ -1,7 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Diagnostics.CodeAnalysis;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
@@ -224,7 +223,7 @@ namespace Microsoft.DotNet.GenAPI
             yield return constructor;
         }
 
-        // Synthesize a base class initializer. 
+        // Synthesize a base class initializer.
         public static ConstructorInitializerSyntax GenerateBaseConstructorInitializer(this IMethodSymbol baseTypeConstructor)
         {
             return SyntaxFactory.ConstructorInitializer(SyntaxKind.BaseConstructorInitializer, baseTypeConstructor.CreateDefaultArgumentList());
@@ -247,34 +246,6 @@ namespace Microsoft.DotNet.GenAPI
             }
 
             return argumentList;
-        }
-
-        // Locates constructor generated by the compiler for `record Foo(...)` syntax
-        // If the type is a record and the compiler generated constructor is found it will be returned, otherwise null.
-        // The compiler will not generate a constructor in the case where the user defined it themself without using an argument list
-        // in the record declaration, or if the record has no parameters.
-        public static bool TryGetRecordConstructor(this INamedTypeSymbol type, [NotNullWhen(true)] out IMethodSymbol? recordConstructor)
-        {
-            if (!type.IsRecord)
-            {
-                recordConstructor = null;
-                return false;
-            }
-
-            // Locate the compiler generated Deconstruct method.
-            var deconstructMethod = (IMethodSymbol?)type.GetMembers("Deconstruct")
-                .FirstOrDefault(m => m is IMethodSymbol && m.IsCompilerGenerated());
-
-            // Locate the compiler generated constructor by matching parameters to Deconstruct - since we cannot locate it with an attribute.
-            recordConstructor = (IMethodSymbol?)type.GetMembers(".ctor")
-                .FirstOrDefault(m => m is IMethodSymbol method &&
-                    method.MethodKind == MethodKind.Constructor &&
-                    (deconstructMethod == null ?
-                        method.Parameters.IsEmpty :
-                        method.Parameters.Select(p => p.Type).SequenceEqual(
-                            deconstructMethod.Parameters.Select(p => p.Type), SymbolEqualityComparer.Default)));
-
-            return recordConstructor != null;
         }
     }
 }
