+++ src/Compatibility/Microsoft.DotNet.ApiSymbolExtensions/SymbolExtensions.cs
@@ -1,6 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
+using System.Diagnostics.CodeAnalysis;
 using Microsoft.CodeAnalysis;
 
 namespace Microsoft.DotNet.ApiSymbolExtensions
@@ -126,5 +127,37 @@ namespace Microsoft.DotNet.ApiSymbolExtensions
             method.MethodKind == MethodKind.EventRemove ||
             method.Name.StartsWith("add_", StringComparison.Ordinal) ||
             method.Name.StartsWith("remove_", StringComparison.Ordinal);
+
+        /// <summary>
+        /// Attempts to locate and return the constructor generated by the compiler for `record Foo(...)` syntax.
+        /// The compiler will not generate a constructor in the case where the user defined it themself without using an argument list
+        /// in the record declaration, or if the record has no parameters.
+        /// </summary>
+        /// <param name="type">The type to check for a compiler generated constructor.</param>
+        /// <param name="recordConstructor">When this method returns <see langword="true"/>, then the compiler generated constructor for the record is stored in this out parameter, otherwise it becomes <see langword="null" />.</param>
+        /// <returns><see langword="true" /> if the type is a record and the compiler generated constructor is found, otherwise <see langword="false"/>.</returns>
+        public static bool TryGetRecordConstructor(this INamedTypeSymbol type, [NotNullWhen(true)] out IMethodSymbol? recordConstructor)
+        {
+            if (!type.IsRecord)
+            {
+                recordConstructor = null;
+                return false;
+            }
+
+            // Locate the compiler generated Deconstruct method.
+            var deconstructMethod = (IMethodSymbol?)type.GetMembers("Deconstruct")
+                .FirstOrDefault(m => m is IMethodSymbol && m.IsCompilerGenerated());
+
+            // Locate the compiler generated constructor by matching parameters to Deconstruct - since we cannot locate it with an attribute.
+            recordConstructor = (IMethodSymbol?)type.GetMembers(".ctor")
+                .FirstOrDefault(m => m is IMethodSymbol method &&
+                    method.MethodKind == MethodKind.Constructor &&
+                    (deconstructMethod == null ?
+                        method.Parameters.IsEmpty :
+                        method.Parameters.Select(p => p.Type).SequenceEqual(
+                            deconstructMethod.Parameters.Select(p => p.Type), SymbolEqualityComparer.Default)));
+
+            return recordConstructor != null;
+        }
     }
 }
